Enable Users Of ShoppingList To Add Item To List
Attempting to add an item that has no entry in the target shopping
list will result in a new entry being added to the list.

Attemtping to add an item that already has an entry in the target list
will result in the quantity of the existing entry to increase by one.

If the price of the entry is different from the price of the item we
assume that the new item price is correct, therefore we use that.